added import / refactored menu.py
